requirements.txt
openai
arize-phoenix[evals]
numpy==1.23.1
scikit-learn==1.2.0

